requirements.txt
discord.py
os
ctypes
asyncio
atexit
random
math
requests
cogs